+++ fingertree-rm/CHANGELOG.md
@@ -1,5 +1,9 @@
+## Next version
 
-<a id='changelog-0.1.0.0'></a>
+### Breaking changes
+
+* Require a root measure to be a cancellative monoid instead of a group.
+
 ## 0.1.0.0 — 2023-02-16
 
 First release